refactor: rename file to follow kebab-case
