+++ src/app/layout.tsx
@@ -4,7 +4,7 @@ import './globals.css';
 
 import Header from '@/components/header/header';
 import Footer from '@/components/footer/footer';
-import { AuthProvider } from '@/context/AuthContext';
+import { AuthProvider } from '@/context/auth-context';
 const geistSans = Geist({
 	variable: '--font-geist-sans',
 	subsets: ['latin'],
