+++ src/components/header/header.tsx
@@ -10,8 +10,8 @@ import NavBar from './nav-bar';
 
 import { closeIcon, hamburgerIcon } from '@/assets';
 import { desktopNavLinks, mobileNavLinks } from '@/constants/Header/constants';
-import { useIsAuthenticated } from '@/context/AuthContext';
-import { useAuth } from '@/context/AuthContext';
+import { useIsAuthenticated } from '@/context/auth-context';
+import { useAuth } from '@/context/auth-context';
 const Header = () => {
 	const { isAuthenticated, loading } = useIsAuthenticated();
 	const [isMenuOpen, setIsMenuOpen] = useState(false);
