+++ src/components/header/profile-card.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Image from 'next/image';
 import { mockedAvatar } from '@/assets';
-import { useAuth } from '@/context/AuthContext';
+import { useAuth } from '@/context/auth-context';
 import ProfileMenu from '@/components/header/profile-menu';
 import { useState } from 'react';
